+++ lib/src/async_field_base.dart
@@ -71,6 +71,7 @@ class AsyncField<T> {
 
   Object get idKeyAsJson => json.encode(id.key);
 
+  bool _hasValueOrSlate = false;
   T? _value;
 
   String get valueAsJson {
@@ -83,21 +84,11 @@ class AsyncField<T> {
     return isSet ? '$_value' : (defaultValue ?? '?').toString();
   }
 
-  double get valueAsDouble {
-    checkValueTimeout();
-    return double.parse(valueAsString);
-  }
+  double get valueAsDouble => double.parse(valueAsString);
 
-  int get valueAsInt {
-    checkValueTimeout();
-    return int.parse(valueAsString);
-  }
+  int get valueAsInt => int.parse(valueAsString);
 
-  bool get valueAsBool {
-    checkValueTimeout();
-    var val = value;
-    return _parseBool(val);
-  }
+  bool get valueAsBool => _parseBool(value);
 
   static bool _parseBool(val) {
     if (val == null) {
@@ -149,8 +140,20 @@ class AsyncField<T> {
     return _value ?? defaultValue;
   }
 
+  /// Returns [value] or a slate version or the [defaultValue].
+  T? get valueOrSlate {
+    var slate = checkValueTimeout();
+    return _value ?? slate ?? defaultValue;
+  }
+
+  /// Returns [value] without [checkValueTimeout].
   T? get valueNoTimeoutCheck {
     return _value ?? defaultValue;
+  }
+
+  /// Returns [valueOrSlate] without [checkValueTimeout].
+  T? get valueOrSlateNoTimeoutCheck {
+    return _value ?? _slateValue ?? defaultValue;
   }
 
   /// Same as [valueTime], but `millisecondsSinceEpoch`.
@@ -180,29 +183,44 @@ class AsyncField<T> {
 
   /// Returns `true` if this [value] is valid, based in the [timeout] and [isSet].
   bool get isValid {
+    if (!isSet || isSlate) return false;
+
     var timeout = this.timeout;
-    return timeout != null
-        ? isSet && valueElapsedTime <= timeout.inMilliseconds
-        : isSet;
+    if (timeout != null) {
+      return valueElapsedTime <= timeout.inMilliseconds;
+    } else {
+      return true;
+    }
   }
 
   /// Returns `true` if this value is expired, based in the [timeout].
   bool get isExpire {
     var timeout = this.timeout;
-
     return timeout != null && valueElapsedTime > timeout.inMilliseconds;
   }
 
-  /// Returns `true` if this field value is set.
-  bool get isSet => _valueTime != null;
+  /// Returns `true` if this field value is set and NOT slate.
+  bool get isSet => _hasValueOrSlate && _slateValue == null;
+
+  /// Returns `true` if this field value is set OR have a slate value.
+  bool get isSetOrSlate => _hasValueOrSlate;
+
+  /// Returns `true` if this field value has only a slate version.
+  bool get isSlate => _hasValueOrSlate && _slateValue != null;
+
+  T? _slateValue;
 
   /// Checks [value] [timeout] and invalidate it if [isExpire].
   T? checkValueTimeout() {
     if (isExpire) {
-      var slate = _value;
-      _value = null;
-      _valueTime = null;
-      return slate;
+      if (_value != null) {
+        var slate = _value;
+        _slateValue = slate;
+        _value = null;
+        return slate;
+      } else {
+        return _slateValue;
+      }
     } else {
       return null;
     }
@@ -212,7 +230,10 @@ class AsyncField<T> {
   T? disposeValue() {
     var slate = _value;
     _value = null;
+    _hasValueOrSlate = false;
     _valueTime = null;
+    _slateValue = null;
+    _prevValue = null;
     return slate;
   }
 
@@ -280,18 +301,24 @@ class AsyncField<T> {
   /// Filters values changes before trigger [onChange].
   bool Function(T? prevValue, T? value)? onChangeFilter;
 
-  bool _isChangingValue(T? prevValue, T? value) {
-    final changeValidator = this.onChangeFilter;
-    if (changeValidator != null) {
-      return changeValidator(prevValue, value);
+  bool _canNotifyChange(T? prevValue, T? value) {
+    final onChangeFilter = this.onChangeFilter;
+    if (onChangeFilter != null) {
+      return onChangeFilter(prevValue, value);
     } else {
       return true;
     }
   }
 
+  // Temporary holds the previous value while [set] and [_set]
+  T? _prevValue;
+
   /// Sets this field with [value].
   FutureOr<T> set(T value) {
+    _prevValue = _value ?? _slateValue;
+
     _value = value;
+    _hasValueOrSlate = true;
     _valueTime = DateTime.now().millisecondsSinceEpoch;
 
     return save();
@@ -299,12 +326,15 @@ class AsyncField<T> {
 
   // Internal _set called by [_onSave] and [onFetch].
   AsyncField<T> _set(T value) {
-    var prevValue = _value;
+    var prevValue = _prevValue;
 
     _value = value;
+    _hasValueOrSlate = true;
     _valueTime = DateTime.now().millisecondsSinceEpoch;
+    _slateValue = null;
+    _prevValue = null;
 
-    if (_isChangingValue(prevValue, value)) {
+    if (_canNotifyChange(prevValue, value)) {
       _onChangeController.add(this);
     }
 
@@ -323,6 +353,7 @@ class AsyncField<T> {
         periodicRefresh.inMilliseconds - valueElapsedTime;
 
     if (timeToPeriodicRefresh <= 0) {
+      // ignore: discarded_futures
       refresh();
       return;
     }
@@ -334,6 +365,7 @@ class AsyncField<T> {
       if (valueElapsedTime < periodicRefresh.inMilliseconds) {
         _checkPeriodicRefresh();
       } else {
+        // ignore: discarded_futures
         refresh();
       }
     });
@@ -382,6 +414,8 @@ class AsyncField<T> {
             "Closed `storage`: can't return `null` value as `$T` @ $runtimeType#${id.key}");
       }
     }
+
+    _prevValue = _value ?? _slateValue;
 
     var fetcher = this.fetcher;
 
@@ -484,6 +518,9 @@ class AsyncField<T> {
       _deletedValue = _value;
 
       _value = null;
+      _slateValue = null;
+      _prevValue = null;
+      _hasValueOrSlate = false;
       _valueTime = DateTime.now().millisecondsSinceEpoch;
 
       _onDeleteController.add(this);
@@ -558,12 +595,31 @@ class AsyncField<T> {
     }
   }
 
+  /// If `true` allows slate values on [toDSXValue]. Default: `false`
+  bool dsxValueAllowSlate = false;
+
+  /// If `true` allows automatic fetch on [toDSXValue]. Default: `true`
+  bool dsxValueAllowAutoFetch = true;
+
   /// DSX dynamic interface (package `dom_builder`).
   dynamic toDSXValue() {
-    if (isSet && isValid) {
-      return valueNoTimeoutCheck;
-    } else {
+    if (isSetOrSlate) {
+      if (isValid) {
+        return valueNoTimeoutCheck;
+      } else if (dsxValueAllowSlate) {
+        return valueOrSlateNoTimeoutCheck;
+      }
+    }
+
+    if (dsxValueAllowAutoFetch) {
+      // ignore: discarded_futures
       return get();
+    } else {
+      if (dsxValueAllowSlate) {
+        return valueOrSlateNoTimeoutCheck;
+      } else {
+        return valueNoTimeoutCheck;
+      }
     }
   }